refactor pojograph.external
