+++ core/src/main/scala/com/raphtory/internals/storage/pojograph/entities/external/edge/PojoExEdge.scala
@@ -1,6 +1,7 @@
-package com.raphtory.internals.storage.pojograph.entities.external
+package com.raphtory.internals.storage.pojograph.entities.external.edge
 
 import com.raphtory.internals.storage.pojograph.PojoGraphLens
+import com.raphtory.internals.storage.pojograph.entities.external.PojoExEntity
 import com.raphtory.internals.storage.pojograph.entities.internal.PojoEdge
 
 private[pojograph] class PojoExEdge(
