+++ core/src/main/scala/com/raphtory/internals/storage/pojograph/entities/external/vertex/PojoExplodedVertex.scala
@@ -1,10 +1,8 @@
-package com.raphtory.internals.storage.pojograph.entities.external
+package com.raphtory.internals.storage.pojograph.entities.external.vertex
 
 import com.raphtory.api.analysis.visitor.ExplodedVertex
 import com.raphtory.api.analysis.visitor.HistoricEvent
-import com.raphtory.internals.components.querymanager.FilteredInEdgeMessage
-import com.raphtory.internals.components.querymanager.FilteredOutEdgeMessage
-import com.raphtory.internals.storage.pojograph.messaging.VertexMultiQueue
+import com.raphtory.internals.storage.pojograph.entities.external.edge.PojoExMultilayerEdge
 
 import scala.collection.mutable
 import scala.reflect.ClassTag
