+++ core/src/main/scala/com/raphtory/internals/storage/pojograph/entities/external/vertex/PojoReversedView.scala
@@ -1,4 +1,4 @@
-package com.raphtory.internals.storage.pojograph.entities.external
+package com.raphtory.internals.storage.pojograph.entities.external.vertex
 
 import com.raphtory.api.analysis.visitor.ExplodedVertex
 import com.raphtory.api.analysis.visitor.ReducedVertex
